Start refactoring the tests
